[MS-266] refactor ScanCountdownBar to skip code coverage
